# docs/swagger.yaml
basePath: /api
definitions:
  model.Jamaah:
    properties:
      alamat:
        type: string
      batik_kbih_sudah_diterima:
        type: boolean
      batik_nasional_sudah_dijahit:
        type: boolean
      created_at:
        type: string
      foto_kk:
        type: string
      foto_koper_diterima:
        type: string
      foto_ktp:
        type: string
      foto_paspor:
        type: string
      id:
        type: string
      jenis_kelamin:
        enum:
        - laki-laki
        - perempuan
        type: string
      keterangan:
        type: string
      koper_sudah_diterima:
        type: boolean
      nama:
        type: string
      nik:
        type: string
      no_hp:
        type: string
      no_rekening_haji:
        type: string
      nomor_paspor:
        type: string
      paket:
        enum:
        - haji
        - umroh
        type: string
      pasfoto:
        type: string
      status_pembayaran:
        enum:
        - belum_bayar
        - dp
        - lunas
        type: string
      tanggal_keberangkatan:
        type: string
      tanggal_lahir:
        type: string
      tipe_bank:
        type: string
      updated_at:
        type: string
    required:
    - jenis_kelamin
    - nama
    - nik
    - paket
    - status_pembayaran
    type: object
host: localhost:8080
info:
  contact: {}
  description: API untuk manajemen data jamaah haji dan umroh
  title: Syairullah Haji & Umroh API
  version: "1.0"
paths:
  /jamaah:
    get:
      description: Menampilkan daftar jamaah, bisa difilter berdasarkan paket
      parameters:
      - description: Filter berdasarkan paket
        enum:
        - haji
        - umroh
        in: query
        name: paket
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Jamaah'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List semua jamaah
      tags:
      - jamaah
    post:
      consumes:
      - application/json
      description: Membuat data jamaah haji/umroh baru
      parameters:
      - description: Data jamaah
        in: body
        name: jamaah
        required: true
        schema:
          $ref: '#/definitions/model.Jamaah'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Jamaah'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Tambah jamaah baru
      tags:
      - jamaah
  /jamaah/{id}:
    delete:
      description: Menghapus data jamaah berdasarkan ID
      parameters:
      - description: Jamaah ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Hapus jamaah
      tags:
      - jamaah
    get:
      description: Menampilkan detail jamaah berdasarkan ID
      parameters:
      - description: Jamaah ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Jamaah'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Detail jamaah
      tags:
      - jamaah
    put:
      consumes:
      - application/json
      description: Memperbarui data jamaah berdasarkan ID
      parameters:
      - description: Jamaah ID
        in: path
        name: id
        required: true
        type: string
      - description: Data jamaah
        in: body
        name: jamaah
        required: true
        schema:
          $ref: '#/definitions/model.Jamaah'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update jamaah
      tags:
      - jamaah
  /jamaah/{id}/dokumen/{docType}:
    get:
      description: Download file dokumen jamaah berdasarkan tipe
      parameters:
      - description: Jamaah ID
        in: path
        name: id
        required: true
        type: string
      - description: Tipe dokumen
        enum:
        - ktp
        - kk
        - paspor
        - pasfoto
        - koper_diterima
        in: path
        name: docType
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download dokumen jamaah
      tags:
      - dokumen
  /jamaah/{id}/upload/{docType}:
    post:
      consumes:
      - multipart/form-data
      description: Upload file dokumen (KTP, KK, Paspor, Pasfoto, Koper Diterima)
        untuk jamaah
      parameters:
      - description: Jamaah ID
        in: path
        name: id
        required: true
        type: string
      - description: Tipe dokumen
        enum:
        - ktp
        - kk
        - paspor
        - pasfoto
        - koper_diterima
        in: path
        name: docType
        required: true
        type: string
      - description: File dokumen (jpg, jpeg, png, pdf)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload dokumen jamaah
      tags:
      - dokumen
swagger: "2.0"